- Now the vertex indices (elements) of a Mesh are saved as unsigned ints instead of as unsigned bytes.

+++ src/client/models/3d/mesh.hpp
@@ -63,7 +63,7 @@ class Mesh : public Drawable
         std::vector< glm::vec3 > originalNormals;
 
         // Mesh's triangles.
-        std::vector< std::array< GLubyte, 3 > > triangles;
+        std::vector< std::array< GLuint, 3 > > triangles;
 
         // Mesh's centroid.
         glm::vec4 originalCentroid;
